# docs/api/api.yaml
openapi: 3.0.0
info:
  version: '1'
  title: Social API
  description: >
    The Social API is a comprehensive RESTful interface developed and

    maintained by the Tranvelocity team. It facilitates communication with the

    Social platform, allowing users to access and manage various

    social-related functionalities.


    Authentication and Authorization:


    The API employs a robust authentication and authorization mechanism to

    ensure secure access to resources. To interact with the API, users must

    include the following headers in their requests:


    - `X-Social-Api-Key`: A unique API key assigned to each user for

    authentication purposes.


    - `X-Social-Timestamp`: A Unix timestamp indicating the time of the

    request.


    - `X-Social-Signature`: A cryptographic signature generated using the

    `SHA-512` algorithm based on the API key, timestamp, and secret key. These

    headers are required for all requests.


    - `X-Social-User-SSID`: The User login session ID for user

    authentication. If provided, the API will specify the user's role based on

    this login session.


    - `X-Social-Auth-Id`: The authentication ID providing additional user

    identification.


    The Social API will specify the user's role based on this login session.
    If these `X-Social-User-SSID` and `X-Social-Auth-Id`

    headers are not included, it indicates that the request is sent by

    non-registered users.


    Role-Based Access Control:


...
servers:
...
paths:
...
  /medias:
    get:
      summary: Retrieve a list of media items
      description: >
        This endpoint allows users to retrieve a list of media items based on
        optional filtering parameters.
      tags:
        - Media-related endpoints
      operationId: get-media-list
      parameters:
        - $ref: '#/components/parameters/X-Social-Api-Key'
        - $ref: '#/components/parameters/X-Social-Timestamp'
        - $ref: '#/components/parameters/X-Social-Signature'
        - $ref: '#/components/parameters/X-Social-User-SSID'
        - $ref: '#/components/parameters/X-Social-Auth-Id'
        - in: query
          name: post_id
          schema:
            type: string
          description: The ID of the post associated with the media items
        - in: query
          name: type
          schema:
            type: string
          description: The type of media items to filter by
      responses:
        '200':
          description: Success response containing a list of media items
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MediaList'
              example:
                data:
                  - id: 01hnwazag8mq5s1ga3zswp6afr
                    post_id: ''
                    type: video
                    path: >-
                      https://123456789.cloudfront.net/social/staging/medias/videos/d93c51e2-6397-4833-afdb-f2c6136b4121_1707126794.mp4
                    thumnail: >-
                      https://123456789.cloudfront.net/social/staging/medias/thumbnail/d93c51e2-6397-4833-afdb-f2c6136b4121_1707126123.png
                    updated_at: '2024-02-05 18:55:59'
                    created_at: '2024-02-05 18:53:14'
                  - id: 01hnwa050qbpffgkpykgmeh012
                    post_id: null
                    type: video
                    path: >-
                      https://123456789.cloudfront.net/social/staging/medias/videos/f5f613a4-62ce-4b91-9814-22faa30c66d9_1707125773.mp4
                    thumnail: >-
                      https://123456789.cloudfront.net/social/staging/medias/thumbnail/d93c51e2-6397-4833-afdb-f2c6136b4121_1707126123.png
                    updated_at: '2024-02-05 18:57:16'
                    created_at: '2024-02-05 18:36:13'
                  - id: 01hnw85ayjxhwzd4affh18mrqr
                    post_id: null
                    type: video
                    path: >-
                      https://123456789.cloudfront.net/social/staging/medias/videos/8890d6d0-95dc-4f39-a43a-61e8016586d3_1707123845.mp4
                    thumnail: >-
                      https://123456789.cloudfront.net/social/staging/medias/thumbnail/d93c51e2-6397-4833-afdb-f2c6136b4121_1707126123.png
                    updated_at: '2024-02-05 18:05:33'
                    created_at: '2024-02-05 18:04:06'
                success: true
                code: 200
                pagination:
                  offset: 0
                  limit: 30
                  total: 3
        '401':
          $ref: '#/components/responses/401AuthenticationFailedResponse'
        '403':
          $ref: '#/components/responses/403ForbiddenResponse'
    post:
      operationId: create-a-new-media
      summary: Create a new media item
      description: >
        This endpoint allows users to upload media files such as images and
        videos. If the `post_id` is null, it indicates that the media is being
        uploaded independently and will not be associated with any specific
        post, similar to the functionality on platforms like Facebook or
        Twitter.

        - Supported file formats for images include APNG, PNG, JPEG, and JPG.

        - Supported formats for videos include MP4, MOV, WMV, and WAV.

        - Uploaded files are subject to size limitations:
          - Images must not exceed 10MB.
          - Videos must not exceed 1GB.
      tags:
        - Media-related endpoints
      parameters:
        - $ref: '#/components/parameters/X-Social-Api-Key'
        - $ref: '#/components/parameters/X-Social-Timestamp'
        - $ref: '#/components/parameters/X-Social-Signature'
        - $ref: '#/components/parameters/X-Social-User-SSID'
        - $ref: '#/components/parameters/X-Social-Auth-Id'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: form-data
                thumbnail:
                  type: string
                  format: form-data
              required:
                - file
      responses:
        '201':
          description: Success response indicating the media item was successfully created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MediaItem'
              example:
                data:
                  id: 01hp0zsjgzhyetx9d605fd3h7q
                  post_id: null
                  type: image
                  path: >-
                    https://123456789.cloudfront.net/social/staging/medias//images/e0c8535a-85a8-4225-9b61-bba6d832708e_1707282842.jpg
                  thumnail: >-
                    https://123456789.cloudfront.net/social/staging/medias/thumbnail/d93c51e2-6397-4833-afdb-f2c6136b4121_1707126123.png
                  updated_at: '2024-02-07 14:14:04'
                  created_at: '2024-02-07 14:14:04'
                success: true
                code: 201
        '400':
          description: >-
            Bad request response indicating validation errors in the request
            parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorWithValidation'
              example:
                data: []
                success: false
                code: 400
                errors:
                  message: >-
                    The request parameters are incorrect, please make sure to
                    follow the documentation about request parameters of the
                    resource.
                  validation:
                    - attribute: file
                      errors:
                        - key: Required
                          message: required|The file field is required.
        '401':
          $ref: '#/components/responses/401AuthenticationFailedResponse'
        '403':
          $ref: '#/components/responses/403ForbiddenResponse'
        '500':
          description: >-
            Internal server error response indicating a failure occurred during
            the file upload process
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                success: false
                code: 500
                errors:
                  message: >-
                    Failed to upload file. The S3 file path is
                    social/local/medias/images/66ef66a6-1ce0-4593-a424-98a339308a92_1707283158.jpg
                  code: 500002
  '/medias/{id}':
    get:
      summary: Retrieve a specific media item
      description: |
        This endpoint allows users to retrieve a specific media item by its ID.
      tags:
        - Media-related endpoints
      operationId: retrieve-a-specific-media
      parameters:
        - $ref: '#/components/parameters/X-Social-Api-Key'
        - $ref: '#/components/parameters/X-Social-Timestamp'
        - $ref: '#/components/parameters/X-Social-Signature'
        - $ref: '#/components/parameters/X-Social-User-SSID'
        - $ref: '#/components/parameters/X-Social-Auth-Id'
        - in: path
          name: id
          required: true
          schema:
            type: string
            example: 01hnwazag8mq5s1ga3zswp6afr
          description: The ID of the media item to retrieve
      responses:
        '200':
          description: Success response containing the requested media item
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MediaItem'
              example:
                data:
                  id: 01hnwazag8mq5s1ga3zswp6afr
                  post_id: 01hnw7nqm19g9302c7wq391nvm
                  type: video
                  path: >-
                    https://123456789.cloudfront.net/social/staging/medias/videos/d93c51e2-6397-4833-afdb-f2c6136b4121_1707126794.mp4
                  thumnail: >-
                    https://123456789.cloudfront.net/social/staging/medias/thumbnail/d93c51e2-6397-4833-afdb-f2c6136b4121_1707126123.png
                  updated_at: '2024-02-05 18:55:59'
                  created_at: '2024-02-05 18:53:14'
                success: true
                code: 200
                pagination: null
        '401':
          $ref: '#/components/responses/401AuthenticationFailedResponse'
        '403':
          $ref: '#/components/responses/403ForbiddenResponse'
        '404':
          description: >-
            Not found response indicating the requested media item does not
            exist
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                success: false
                code: 404
                errors:
                  message: ID 01hnwazag8mq5s1ga3zswp6afr1に一致するメディアは見つかりませんでした。
                  code: 404001
    put:
      summary: Update a media
      description: Update an existing media with the provided data.
      tags:
        - Media-related endpoints
      operationId: update-a-specific-media
      parameters:
        - $ref: '#/components/parameters/X-Social-Api-Key'
        - $ref: '#/components/parameters/X-Social-Timestamp'
        - $ref: '#/components/parameters/X-Social-Signature'
        - $ref: '#/components/parameters/X-Social-User-SSID'
        - $ref: '#/components/parameters/X-Social-Auth-Id'
        - name: id
          in: path
          description: The ID of the media to be updated
          required: true
          schema:
            type: string
            example: 01hnwazag8mq5s1ga3zswp6afr
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                post_id:
                  type: string
                  example: 01hnwazag8mq5s1ga3zswp6afr
              required:
                - post_id
      responses:
        '200':
          description: The post was successfully updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SinglePostItemResponse'
              example:
                data:
                  id: 01hnwazag8mq5s1ga3zswp6afr
                  post_id: 01hnw7pynwbmqe5rghta5s1ed6
                  type: video
                  path: >-
                    https://123456789.cloudfront.net/social/staging/medias/videos/d93c51e2-6397-4833-afdb-f2c6136b4121_1707126794.mp4
                  thumnail: >-
                    https://123456789.cloudfront.net/social/staging/medias/thumbnail/d93c51e2-6397-4833-afdb-f2c6136b4121_1707126123.png
                  updated_at: '2024-02-07 16:08:43'
                  created_at: '2024-02-05 18:53:14'
                success: true
                code: 200
        '401':
          $ref: '#/components/responses/401AuthenticationFailedResponse'
        '403':
          $ref: '#/components/responses/403ForbiddenResponse'
        '404':
          description: The requested post or media was not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                example-1:
                  value:
                    success: false
                    code: 404
                    errors:
                      message: ID 01hnwazag8mq5s1ga3zswp6afr1に一致するメディアは見つかりませんでした。
                      code: 404001
                example-2:
                  value:
                    success: false
                    code: 404
                    errors:
                      message: ID 01hnw7nqm19g9302c7wq391nvm11に一致する投稿は見つかりませんでした。
                      code: 404001
        '409':
          description: Conflict - Unable to associate media with the specified post
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                success: false
                code: 409
                errors:
                  message: >-
                    メディア（ID
                    01hnwazag8mq5s1ga3zswp6afr）は該当の投稿に関連付けることができませんでした。既に別の投稿に関連付けられています。
                  code: 409001
    delete:
      summary: Delete a media
      description: Delete the specified media resource.
      tags:
        - Media-related endpoints
      operationId: delete-a-specific-media
      parameters:
        - $ref: '#/components/parameters/X-Social-Api-Key'
        - $ref: '#/components/parameters/X-Social-Timestamp'
        - $ref: '#/components/parameters/X-Social-Signature'
        - $ref: '#/components/parameters/X-Social-User-SSID'
        - $ref: '#/components/parameters/X-Social-Auth-Id'
        - name: id
          in: path
          description: The ID of the media to be updated
          required: true
          schema:
            type: string
            example: 01hnwazag8mq5s1ga3zswp6afr
      responses:
        '204':
          description: The media was successfully deleted
          content:
            application/json:
              schema:
                type: string
              examples:
                no_content:
                  value: ''
        '401':
          $ref: '#/components/responses/401AuthenticationFailedResponse'
        '403':
          $ref: '#/components/responses/403ForbiddenResponse'
        '404':
          description: The requested media was not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                success: false
                code: 404
                errors:
                  message: ID 01hnwa050qbpffgkpykgmeh012に一致するメディアは見つかりませんでした。
                  code: 404001
...
components:
  parameters:
...
    X-Social-Api-Key:
      name: X-Social-Api-Key
      in: header
      description: |
        An unique string provided by Tranvelocity when you create an account. You
        must include this API key in the specified HTTP request header for every
        API request to authenticate your requests.
      required: true
      schema:
        type: string
        example: FSFSDFSDFDSpHYv7kMeb0yVkn2343242L091nI5GLkKPfAt234324bmNEaR
    X-Social-Timestamp:
      name: X-Social-Timestamp
      in: header
      description: |
        An Unix epoch value representing the current time. You must include this
        timestamp in the specified HTTP request header for every API request to
        prevent replay attacks and ensure request freshness.
      required: true
      schema:
        type: integer
        example: 1595907741
    X-Social-Signature:
      name: X-Social-Signature
      in: header
      description: >
        A signature that is HMAC-encoded with SHA-512 digest. You must include
        this

        signature in the specified HTTP request header for every API request to

        ensure data integrity and authenticity.
      required: true
      schema:
        type: string
        example: 5d714ce2385d1cf96260806a50f3f7e68b928cc8
    X-Social-User-SSID:
      name: X-Social-User-SSID
      in: header
      description: |
        The session ID for User.
      required: true
      schema:
        type: string
        example: MBhdHsd4ckmKNNDaScHCJjuPY5pgGFQOhMSSqPis
    X-Social-Auth-Id:
      name: X-Social-Auth-Id
      in: header
      description: |
        The ID used for authentication.
      required: true
      schema:
        type: string
        example: ebbae0f297e50fa92f2eaaa340a1f9125cad6e2d
...
  schemas:
...
    Comment:
      type: object
      properties:
        id:
          type: string
          description: The unique identifier of the comment
        user_id:
          type: integer
          nullable: true
          description: The unique identifier of the comment in User
        post_id:
          type: string
          description: The ID of the post associated with the comment
        comment:
          type: string
          description: The content of the comment
        is_hidden:
          type: boolean
          description: Indicates whether the comment is hidden
        updated_at:
          type: string
          format: date-time
          description: The timestamp when the comment was last updated
        created_at:
          type: string
          format: date-time
          description: The timestamp when the comment was created
...
    ErrorWithValidation:
      type: object
      properties:
        data:
          type: array
          items: {}
          description: An empty array indicating no data is returned
        success:
          type: boolean
        code:
          type: integer
        errors:
          type: object
          properties:
            message:
              type: string
            validation:
              type: array
              items:
                type: object
                properties:
                  attribute:
                    type: string
                  errors:
                    type: array
                    items:
                      type: object
                      properties:
                        key:
                          type: string
                        message:
                          type: string
    MediaList:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/MediaItem'
        success:
          type: boolean
        code:
          type: integer
        pagination:
          $ref: '#/components/schemas/Pagination'
    MediaItem:
      type: object
      properties:
        id:
          type: string
          description: The unique identifier of the media item
        post_id:
          type: string
          nullable: true
          description: The ID of the post associated with the media item
        type:
          type: string
          description: The type of the media item
        path:
          type: string
          description: The path to the media file
        updated_at:
          type: string
          format: date-time
          description: The date and time when the media item was last updated
        created_at:
          type: string
          format: date-time
          description: The date and time when the media item was created
    Pagination:
      type: object
      properties:
        offset:
          type: integer
          description: The offset of the current page in pagination
          example: 0
        limit:
          type: integer
          description: The maximum number of items per page
          example: 30
        total:
          type: integer
          description: The total number of items available
          example: 1
    SinglePostItemResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            id:
              type: string
              description: The unique identifier of the post
              example: 01hqt819gw8q83hfedkpbq5cp0
            admin_uuid:
              type: string
              description: The UUID of the admin who created the post
              example: 6c3a77e6-06cb-43f5-bb49-3b44ffc0eaaa
            type:
              type: string
              description: The type of post
              example: premium
            content:
              type: string
              description: The content of the post
              example: Test content here
            is_published:
              type: boolean
              description: Indicates whether the post is published
              example: true
            published_start_at:
              type: string
              format: date-time
              description: The start date and time of the post's publication
              example: '2024-09-01 00:00:00'
            published_end_at:
              type: string
              format: date-time
              description: The end date and time of the post's publication
              example: '2025-12-01 18:00:00'
            updated_at:
              type: string
              format: date-time
              description: The date and time when the post was last updated
              example: '2024-02-29 19:54:48'
            created_at:
              type: string
              format: date-time
              description: The date and time when the post was created
              example: '2024-02-29 19:54:48'
            total_medias:
              type: integer
              description: The total number of media files attached to the post
              example: 4
            medias:
              type: object
              properties:
                images:
                  type: array
                  items:
                    type: object
                    format: uri
                    description: URL of the image media files
                  example:
                    - src: >-
                        https://dev-social.api.com/medias/images/7e30d400-7997-426e-b4f6-5b0e3e5082fc_1707123584.jpg
                      id: 01hr6d05ftx6hptm2pw2mbrnb2
                    - src: >-
                        https://dev-social.api.com/medias/images/7e30d400-7997-426e-b4f6-5b0e3e5082fc_1707123642.jpg
                      id: 01hr6d05ftx6hptm2pw2mbrg432
                videos:
                  type: array
                  items:
                    type: object
                    format: uri
                    description: URL of the video media files
                  example:
                    - thumbnail: >-
                        https://123456789.cloudfront.net/social/local/medias/thumbnails/97f99721-1615-4572-a612-017a3f193b11.png
                      src: >-
                        https://123456789.cloudfront.net/social/local/video/3d6d2109-a6c9-4b27-a018-c43e389c0567.mp4
                      id: f9c0db9e6125cb1829b1
                    - thumbnail: >-
                        https://123456789.cloudfront.net/social/local/medias/thumbnails/97f99721-1615-4572-a612-017a3f193b11.png
                      src: >-
                        https://123456789.cloudfront.net/social/local/video/c63a6455-54e0-48af-9191-7d454fab39ae.mov
                      id: f9c0db9e6125cb1829b1
            comments:
              type: array
              description: List of comments associated with the post
              items:
                $ref: '#/components/schemas/Comment'
            poster:
              $ref: '#/components/schemas/Poster'
        success:
          type: boolean
          description: Indicates whether the request was successful
          example: true
        code:
          type: integer
          description: The status code of the response
          example: 200
      example:
        data:
          id: 01hqt819gw8q83hfedkpbq5cp0
          admin_uuid: 6c3a77e6-06cb-43f5-bb49-3b44ffc0eaaa
          type: premium
          content: Test content here
          is_published: true
          published_start_at: '2024-09-01 00:00:00'
          published_end_at: '2025-12-01 18:00:00'
          updated_at: '2024-02-29 19:54:48'
          created_at: '2024-02-29 19:54:48'
          total_medias: 4
          medias:
            images:
              - src: >-
                  https://dev-social.api.com/medias/images/7e30d400-7997-426e-b4f6-5b0e3e5082fc_1707123584.jpg
                id: 01hr6d05ftx6hptm2pw2mbrnb2
              - src: >-
                  https://dev-social.api.com/medias/images/7e30d400-7997-426e-b4f6-5b0e3e5082fc_1707123642.jpg
                id: 01hr6d05ftx6hptm2pw2mbrg432
            videos:
              - thumbnail: >-
                  https://123456789.cloudfront.net/social/local/medias/thumbnails/97f99721-1615-4572-a612-017a3f193b11.png
                src: >-
                  https://123456789.cloudfront.net/social/local/video/3d6d2109-a6c9-4b27-a018-c43e389c0567.mp4
                id: f9c0db9e6125cb1829b1
              - thumbnail: >-
                  https://123456789.cloudfront.net/social/local/medias/thumbnails/97f99721-1615-4572-a612-017a3f193b11.png
                src: >-
                  https://123456789.cloudfront.net/social/local/video/c63a6455-54e0-48af-9191-7d454fab39ae.mov
                id: f9c0db9e6125cb1829b1
          comments: []
          poster:
            id: 01hqszfkc6fpfya5m7n42zpd34
            admin_uuid: 6c3a77e6-06cb-43f5-bb49-3b44ffc0eaaa
            user_id: 16614
            nickname: hoctran
            avatar: >-
              https://stage.user.com/file/fileconsumer/doo/selector/downloader/phase/download_file/type/images/file_id/phpv2w4cF
            description: hoc tran .
            updated_at: '2024-02-29 17:25:20'
            created_at: '2024-02-29 17:25:20'
        success: true
        code: 200
...
    Error:
      type: object
      properties:
        success:
          type: boolean
          description: Indicates whether the request was successful
        code:
          type: integer
          description: The status code of the error
        errors:
          type: object
          properties:
            message:
              type: string
              description: A descriptive message explaining the error
            code:
              type: integer
              description: A unique error code identifying the specific error
  responses:
    401AuthenticationFailedResponse:
      description: Unauthorized. HMAC authentication failed or invalid session ID.
      content:
        application/json:
          examples:
            HMACAuthenticationFailed:
              summary: HMAC authentication failed.
              value:
                success: false
                code: 401
                errors:
                  message: >-
                    HMAC authentication was failed. Please make sure your
                    credentials again.
                  code: 401001
            InvalidSessionID:
              summary: Invalid session ID.
              value:
                success: false
                code: 401
                errors:
                  message: Invalid session ID.
                  code: 401004
    403ForbiddenResponse:
      description: Forbidden - Access Denied
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                description: Indicates whether the request was successful
                example: false
              code:
                type: integer
                description: The HTTP status code of the response
                example: 403
              errors:
                type: object
                properties:
                  message:
                    type: string
                    description: A message explaining the reason for the forbidden access
                    example: 指定されたデータをいじる権限がありません。
                  code:
                    type: integer
                    description: A code indicating the specific error
                    example: 403001
...